button up the build
* write the temp git hash file in xcompile
* fix tokio_unstable flag
* push Bazel built images to Docker
